Added custom task queues to the rmq thread comm
can now do:
task_queue = comm.task_queue('my_task_queue')
task_queue.add_task_subscriber(lambda _, task: print(task))
task_queue.task_send('Do this!')
