+++ kiwipy/rmq/__init__.py
@@ -1,6 +1,8 @@
 # pylint: disable=wildcard-import
 from .communicator import *
 from .tasks import *
+from .threadcomms import *
 from .utils import *
 
-__all__ = (tasks.__all__ + communicator.__all__ + utils.__all__)
+# pylint: disable=undefined-variable
+__all__ = (tasks.__all__ + communicator.__all__ + threadcomms.__all__ + utils.__all__)
